Toggle hide state on clicking tray icon
The previous tray behaviour called `tray.popUpContextMenu()` when
clicked, although this didn't seem to work. The behaviour to open a
context menu on a right click event, though, seems to be handled
implicitly when setting the tray object's context menu object. (Under
Linux, Electron only has a 'click' event for tray icons, unlike Windows
and Mac for which there are explicit 'double-click' and 'right-click'
events.)

This patch adds a `toggleMinimizedToTray()` function that opens the
window if minimized, and minimizes to tray if open; and binds that to
the tray icon's click event. This lets you open (or close) the
mainWindow from the tray icon with a single left-click, in addition to
the existing right-click->show (or Minimize to tray) options in the
context menu.

Some applications--for example Discord--prefer only to have the open
behaviour on a tray click; others--for example Signal--prefer the toggle
behaviour I've implemented here. It would be trivial to change that.